% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astext.R
\name{as_text}
\alias{as_text}
\title{Convert Raw to Text}
\usage{
as_text(x, ...)
}
\arguments{
\item{x}{vector to be converted to text}

\item{...}{parameters passed to \link{readLines} such as \code{encoding} or \code{n}}
}
\description{
Parses a raw vector as lines of text. This is similar to \link{charToRaw} but
splits output by (platform specific) linebreaks and allows for marking
output with a given encoding.
}
\seealso{
\link[base:rawConversion]{base::charToRaw}
}
