% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{sys-deprecated}
\alias{sys-deprecated}
\alias{eval_safe}
\alias{eval_fork}
\title{Deprecated functions}
\usage{
eval_safe(...)

eval_fork(...)
}
\arguments{
\item{...}{see respective functions in the unix package}
}
\description{
These functions have moved into the \code{unix} package. Please update
your references.
}
