% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{eval_safe}
\alias{eval_safe}
\title{Legacy functions}
\usage{
eval_safe(...)
}
\arguments{
\item{...}{see \code{?eval_safe} in unix}
}
\description{
These functions have moved into the \code{unix} package.
}
