% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{exec_with_wait}
\alias{exec_with_wait}
\alias{exec_with_pid}
\title{Execute a program}
\usage{
exec_with_wait(cmd, args = NULL)

exec_with_pid(cmd, args = NULL)
}
\arguments{
\item{cmd}{the command to run}

\item{args}{character vector of arguments to pass}
}
\description{
The \code{exec_with_wait} function runs a program and returns the status code
of the child program. Negative codes indicates the program was terminated
by a SIGNAL. The \code{exec_with_pid} function returns immediately with the PID
of the child program. In this case the state of the child is unknown.
}

