% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy.R
\name{eval_safe}
\alias{eval_safe}
\alias{aa_config}
\title{Legacy functions}
\usage{
eval_safe(...)

aa_config(...)
}
\arguments{
\item{...}{see respective functions in the unix package}
}
\description{
These functions have moved into the \code{unix} package. Please update
your references.
}
