% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{quote}
\alias{quote}
\alias{windows_quote}
\title{Quote arguments on Windows}
\usage{
windows_quote(args)
}
\arguments{
\item{args}{character vector with arguments}
}
\description{
Properly escapes spaces and double quotes in shell arguments so that they get
properly parsed by most Windows programs.
}
