% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fork.R
\name{eval_fork}
\alias{eval_fork}
\title{Evaluate in fork}
\usage{
eval_fork(expr, envir = parent.frame(), tmp = tempfile("fork"))
}
\arguments{
\item{expr}{expression to evaluate}

\item{envir}{the \link{environment} in which expr is to be evaluated}

\item{tmp}{the value of \link{tempdir} inside the forked process}
}
\description{
Version of \link{eval} which evaluates expression in a temporary fork so
that it has no side effects on the main R session.
Basically a robust version of \code{mccollect(mcparallel(expr))}.
Not available on Windows because it required \code{fork()}.
}
