% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\name{quote}
\alias{quote}
\alias{windows_quote}
\title{Quote arguments on Windows}
\usage{
windows_quote(args)
}
\arguments{
\item{args}{character vector with arguments}
}
\description{
Quotes and escapes shell arguments when needed so that they get properly parsed
by most Windows programs. This function is used internally to automatically quote
system commands, the user should normally not quote arguments manually.
}
\details{
Algorithm is ported to R from
\href{https://github.com/libuv/libuv/blob/v1.23.0/src/win/process.c#L454-L524}{libuv}.
}
